@@ -2,15 +2,21 @@ name: camera_web_integration_tests
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.10.0"
+  sdk: ^3.3.0
+  flutter: ">=3.19.0"
 
 dependencies:
-  camera_platform_interface: ^2.1.0
+  camera_platform_interface: ^2.6.0
   camera_web:
+    # When depending on this package from a real application you should use:
+    #   camera_web: ^x.y.z
+    # See https://dart.dev/tools/pub/dependencies#version-constraints
+    # The example app is bundled with the plugin so we use a path dependency on
+    # the parent directory to use the current plugin's version.
     path: ../
   flutter:
     sdk: flutter
+  web: ^1.0.0
 
 dev_dependencies:
   async: ^2.5.0
